requirements.txt
pandas>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
streamlit>=1.28.0
python-dotenv>=1.0.0
pypdf>=3.17.0
google-search-results>=2.4.2
structlog>=23.2.0
